Trabajos_Practicos/TP06_TDD_Evaluable/backend/requirements.txt
Django>=4.0
djangorestframework
pytest
pytest-django
django-cors-headers